requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.9
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytesseract==0.3.13
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
